test(protocol): move the dependency-contract parity check out of the unit suite
The check reads repository packaging files by design, which belongs in
the protocol suite (repository-contract checks) rather than the
zero-I/O unit suite.
